requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
requests==2.31.0
json-logging==1.3.0